+++ src/onegov.election_day/onegov/election_day/assets/js/tablesorter.js
@@ -1,0 +1,2 @@
+// sort tables wishing to be sorted
+$('table.sortable').tablesorter({widgets: ['staticRow']});